i created new components and i modularized logic to get movies and any components that requires it
